% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{addFactorLevel}
\alias{addFactorLevel}
\title{Add a new level to a factor}
\usage{
addFactorLevel(x, new_level)
}
\arguments{
\item{x}{A factor}

\item{new_level}{A string for new factor level}
}
\value{
A factor
}
\description{
Add a new level to a factor
}
