% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleanData}
\alias{cleanData}
\title{Clean training and test data sets}
\usage{
cleanData(train, test, config_file)
}
\arguments{
\item{train}{A data frame containing the training data}

\item{test}{A data frame containing the test data}

\item{config_file}{A data frame containing the configuration for the data set}
}
\value{
A list containing cleaned train and test data frames
}
\description{
Clean training and test data sets
}
