% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{countNAs}
\alias{countNAs}
\title{Count the number of NAs in each column of a data frame}
\usage{
countNAs(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A vector
}
\description{
Count the number of NAs in each column of a data frame
}
