% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{createConfigFile}
\alias{createConfigFile}
\title{Create configuration file for a data set}
\usage{
createConfigFile(df)
}
\arguments{
\item{df}{A data frame containing the training data set}
}
\value{
A data frame
}
\description{
Create configuration file for a data set
}
