% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{createDateFeatures}
\alias{createDateFeatures}
\title{Create new features based on date fields}
\usage{
createDateFeatures(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
Create new features based on date fields
}
