% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{fitFinalModel}
\alias{fitFinalModel}
\title{Fit all training data using the best model with optimized parameters}
\usage{
fitFinalModel(train_x, train_y, best_fit_model, task_type, metric)
}
\arguments{
\item{train_x}{A data frame containing the training features}

\item{train_y}{A data frame containing the training targets}

\item{best_fit_model}{A list containing the best fit model}

\item{task_type}{A string}

\item{metric}{A string}
}
\value{
A model
}
\description{
Fit all training data using the best model with optimized parameters
}
