% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{fix_factor_levels}
\alias{fix_factor_levels}
\title{Synchronize factor levels between train and test sets}
\usage{
fix_factor_levels(train_x, test_x)
}
\arguments{
\item{train_x}{A data frame containing the training data}

\item{test_x}{A data frame containing the test data}
}
\value{
A list containing train_x and test_x data frames
}
\description{
Synchronize factor levels between train and test sets
}
