% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{fixDataRepresentation}
\alias{fixDataRepresentation}
\title{Change the data representation depending to fit requirements of models}
\usage{
fixDataRepresentation(df, model_name)
}
\arguments{
\item{df}{A data frame}

\item{model_name}{A string}
}
\value{
A data frame
}
\description{
Change the data representation depending to fit requirements of models
}
