% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{getEvaluation}
\alias{getEvaluation}
\title{Predict and evaluate performance of a machine learning model}
\usage{
getEvaluation(ml_model, model_name, valid_x, valid_y, metric)
}
\arguments{
\item{ml_model}{A machine learning model}

\item{model_name}{A string specifying the name of the model}

\item{valid_x}{A data frame containing the features}

\item{valid_y}{A data frame containing the correct answers}

\item{metric}{A string containing the metric to be used}
}
\value{
Score on prediction
}
\description{
Predict and evaluate performance of a machine learning model
}
