% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{getModel}
\alias{getModel}
\title{Get an optimized machine learning model}
\usage{
getModel(train_x, train_y, tast_type, metric, model_name)
}
\arguments{
\item{train_x}{A data frame containing the training features}

\item{train_y}{A data frame containing the training targets}

\item{metric}{A string}

\item{model_name}{A string}

\item{task_type}{A string}
}
\value{
A model object
}
\description{
Get an optimized machine learning model
}
