% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{getModelList}
\alias{getModelList}
\title{Get a list of models to fit}
\usage{
getModelList(task_type)
}
\arguments{
\item{task_type}{A string, "classification" or "regression"}
}
\value{
A vector of models
}
\description{
Get a list of models to fit
}
