% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getModelMatrix}
\alias{getModelMatrix}
\title{Convert a data frame to a model matrix}
\usage{
getModelMatrix(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A model matrix
}
\description{
Convert a data frame to a model matrix
}
