% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getNumClass}
\alias{getNumClass}
\title{Find the number of different classes in labels.}
\usage{
getNumClass(train_y)
}
\arguments{
\item{train_y}{A vector}
}
\value{
A number
}
\description{
Find the number of different classes in labels.
}
