% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{getPartitionIndex}
\alias{getPartitionIndex}
\title{Get partition index for training and validation sets}
\usage{
getPartitionIndex(target_col)
}
\arguments{
\item{target_col}{A vector}
}
\value{
A vector
}
\description{
Get partition index for training and validation sets
}
