% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{getPreProcFactor}
\alias{getPreProcFactor}
\title{Preprocess all factor columns of a data frame}
\usage{
getPreProcFactor(dt)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
Preprocess all factor columns of a data frame
}
