% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{getPreProcNumeric}
\alias{getPreProcNumeric}
\title{Preprocess all numeric columns of a data frame}
\usage{
getPreProcNumeric(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A function
}
\description{
Preprocess all numeric columns of a data frame
}
