% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getSingleValueCols}
\alias{getSingleValueCols}
\title{Find data frame columns where all values are the same}
\usage{
getSingleValueCols(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A character vector
}
\description{
Find data frame columns where all values are the same
}
