% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getSparseMatrix}
\alias{getSparseMatrix}
\title{Convert a data frame to a sparse model matrix}
\usage{
getSparseMatrix(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A sparse model matrix
}
\description{
Convert a data frame to a sparse model matrix
}
