% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{getSubmission}
\alias{getSubmission}
\title{Create submission data frame}
\usage{
getSubmission(final_model, model_name, test_x, Id)
}
\arguments{
\item{final_model}{A machine learning model}

\item{model_name}{A string}

\item{test_x}{A data frame containing the test features}

\item{Id}{A vector containing the submission Ids}
}
\value{
A data frame
}
\description{
Create submission data frame
}
