% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{getTaskParams}
\alias{getTaskParams}
\title{Get task specific parameters}
\usage{
getTaskParams(task_type, train_y)
}
\arguments{
\item{task_type}{A string}

\item{train_y}{A vector}
}
\value{
A list
}
\description{
Get task specific parameters
}
