% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_identifier.R
\name{identifyTask}
\alias{identifyTask}
\title{Identify the type of the learning task}
\usage{
identifyTask(target_col)
}
\arguments{
\item{target_col}{A vector}
}
\value{
A string
}
\description{
Identify the type of the learning task
}
