% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imputeDefault}
\alias{imputeDefault}
\title{Impute missing values with default values}
\usage{
imputeDefault(x)
}
\arguments{
\item{x}{A factor, character, or numeric vector}
}
\value{
A vector
}
\description{
Impute missing values with default values
}
