% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imputeNA}
\alias{imputeNA}
\title{Impute missing values in a column or vector}
\usage{
imputeNA(x, impute = NA, impute_value = NULL)
}
\arguments{
\item{x}{A vector}

\item{impute}{A string specifying imputation method}

\item{impute_value}{Constant value for replacing NAs}
}
\value{
A vector
}
\description{
Impute missing values in a column or vector
}
