% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imputeNAs}
\alias{imputeNAs}
\title{Impute missing values in train or test data set}
\usage{
imputeNAs(df, config_file)
}
\arguments{
\item{df}{A data frame}

\item{config_file}{A data frame containing the configuration data}
}
\value{
A data frame
}
\description{
Impute missing values in train or test data set
}
