% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{imputeValue}
\alias{imputeValue}
\title{Replace NAs with a specific value}
\usage{
imputeValue(x, impute_value)
}
\arguments{
\item{x}{A vector}

\item{impute_value}{A value for replacing NAs}
}
\value{
A vector
}
\description{
Replace NAs with a specific value
}
