% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{isBetter}
\alias{isBetter}
\title{Check if the new score is better than the current best score}
\usage{
isBetter(new_score, best_score, minimize_score)
}
\arguments{
\item{new_score}{A number}

\item{best_score}{A number}

\item{minimize}{A boolean}
}
\value{
A boolean
}
\description{
Check if the new score is better than the current best score
}
