% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{processText}
\alias{processText}
\title{Preprocess all character columns of a data frame}
\usage{
processText(dt)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
Preprocess all character columns of a data frame
}
