% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{selectModel}
\alias{selectModel}
\title{Select the best performing model}
\usage{
selectModel(train_x, train_y, valid_x, valid_y, task_type, metric,
  minimize_score = TRUE, best_fit_model = NULL, model_list = NULL,
  verbose = TRUE)
}
\arguments{
\item{train_x}{A data frame for training features}

\item{train_y}{A data frame for training targets}

\item{valid_x}{A data frame for validation features}

\item{valid_y}{A data frame for validation targets}

\item{metric}{A string for metric type}

\item{minimize_score}{A boolean}

\item{best_fit_model}{An optional list containing an earlier best model}

\item{model_list}{An optional vector of models to fit}

\item{verbose}{A boolean}
}
\value{
A list containing the best fit model
}
\description{
Select the best performing model
}
