% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitter.R
\name{splitData}
\alias{splitData}
\title{Split data to training and validation sets}
\usage{
splitData(df, target_col)
}
\arguments{
\item{df}{A data frame}

\item{target_col}{A string}
}
\value{
A list of data frames
}
\description{
Split data to training and validation sets
}
