% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{update_classes}
\alias{update_classes}
\title{Change classes in a data frame according to configuration file}
\usage{
update_classes(data, config_file)
}
\arguments{
\item{data}{A data frame}

\item{config_file}{A data frame containing the configuration data}
}
\value{
A data frame
}
\description{
Change classes in a data frame according to configuration file
}
