% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selector.R
\name{updateBestModel}
\alias{updateBestModel}
\title{Update best_fit_model if the new model is better than the current best model}
\usage{
updateBestModel(new_score, new_score_train, new_model_name, new_model,
  best_fit_model, minimize_score)
}
\arguments{
\item{new_score}{A number}

\item{new_score_train}{A number}

\item{new_model_name}{A string}

\item{new_model}{A model}

\item{best_fit_model}{A list containing the current best model}

\item{minimize_score}{A boolean}
}
\value{
A list containing the best fit model
}
\description{
Update best_fit_model if the new model is better than the current best model
}
